+++ front/src/index.js
@@ -3,12 +3,15 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import { BrowserRouter } from 'react-router-dom';
+import { CurrentUserProvider } from './context/currentUser-context';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
 	<React.StrictMode>
 		<BrowserRouter>
-			<App />
+			<CurrentUserProvider>
+				<App />
+			</CurrentUserProvider>
 		</BrowserRouter>
 	</React.StrictMode>,
 );
